requirements.txt
yt-dlp
python-dotenv
google-api-python-client>=2.100.0
google-auth>=2.22.0
pytest>=7.4.0
pytest-mock>=3.11.0
